servo_xpath

Trait Node

pub trait Node: Clone + Debug {
    type Context;
    type ProcessingInstruction: ProcessingInstruction;
    type Document: Document<Node = Self>;
    type Attribute: Attribute<Node = Self>;
    type Element: Element<Node = Self, Context = Self::Context>;
    type Opaque: Eq + Hash + 'static;

    // Required methods
    fn is_comment(&self) -> bool;
    fn is_text(&self) -> bool;
    fn text_content(&self) -> String;
    fn language(&self) -> Option<String>;
    fn parent(&self) -> Option<Self>;
    fn children(&self) -> impl Iterator<Item = Self>;
    fn compare_tree_order(
        &self,
        cx: &mut Self::Context,
        other: &Self,
    ) -> Ordering;
    fn traverse_preorder(&self) -> impl Iterator<Item = Self>;
    fn inclusive_ancestors(&self) -> impl Iterator<Item = Self>;
    fn preceding_nodes(&self) -> impl Iterator<Item = Self>;
    fn following_nodes(&self) -> impl Iterator<Item = Self>;
    fn preceding_siblings(&self) -> impl Iterator<Item = Self>;
    fn following_siblings(&self) -> impl Iterator<Item = Self>;
    fn owner_document(&self) -> Self::Document;
    fn to_opaque(&self) -> Self::Opaque;
    fn as_processing_instruction(&self) -> Option<Self::ProcessingInstruction>;
    fn as_attribute(&self) -> Option<Self::Attribute>;
    fn as_element(&self) -> Option<Self::Element>;
    fn get_root_node(&self) -> Self;

}

A handle to a DOM node exposing all functionality needed by xpath.

Required Associated Types

type Context

type ProcessingInstruction: ProcessingInstruction

type Document: Document<Node = Self>

type Attribute: Attribute<Node = Self>

type Element: Element<Node = Self, Context = Self::Context>

type Opaque: Eq + Hash + 'static

Required Methods

fn is_comment(&self) -> bool

fn is_text(&self) -> bool

fn text_content(&self) -> String

Equivalent to textContent attribute.

fn language(&self) -> Option<String>

https://html.spec.whatwg.org/multipage/#language

fn parent(&self) -> Option<Self>

fn children(&self) -> impl Iterator<Item = Self>

fn compare_tree_order(&self, cx: &mut Self::Context, other: &Self) -> Ordering

https://dom.spec.whatwg.org/#concept-tree-order

fn traverse_preorder(&self) -> impl Iterator<Item = Self>

A non-shadow-including preorder traversal.

fn inclusive_ancestors(&self) -> impl Iterator<Item = Self>

fn preceding_nodes(&self) -> impl Iterator<Item = Self>

Return an iterator over all nodes that come before self in tree order, excluding any ancestors and attribute nodes.

fn following_nodes(&self) -> impl Iterator<Item = Self>

Return an iterator over all nodes that come after self in tree order, excluding any descendants and attribute nodes.

fn preceding_siblings(&self) -> impl Iterator<Item = Self>

fn following_siblings(&self) -> impl Iterator<Item = Self>

fn owner_document(&self) -> Self::Document

fn to_opaque(&self) -> Self::Opaque

fn as_processing_instruction(&self) -> Option<Self::ProcessingInstruction>

fn as_attribute(&self) -> Option<Self::Attribute>

fn as_element(&self) -> Option<Self::Element>

fn get_root_node(&self) -> Self

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.